+++ packages/web/src/session/constants.ts
@@ -15,23 +15,8 @@
  * limitations under the License.
  *
  */
-
-import * as assert from 'assert'
-import { generateId } from '../src/utils'
-import { findCookieValue } from '../src/session/cookie-session'
-
-describe('generateId', () => {
-	it('should generate IDs of 64 and 128 bits', () => {
-		const id64 = generateId(64)
-		const id128 = generateId(128)
-		assert.strictEqual(id64.length, 16)
-		assert.strictEqual(id128.length, 32)
-		assert.ok(id64.match('^[0-9a-z]+$'))
-		assert.ok(id128.match('^[0-9a-z]+$'))
-	})
-})
-describe('findCookieValue', () => {
-	it('should not find unset cookie', () => {
-		assert.ok(findCookieValue('nosuchCookie') === undefined)
-	})
-})
+export const SESSION_ID_LENGTH = 32
+export const SESSION_DURATION_SECONDS = 4 * 60 * 60 // 4 hours
+export const SESSION_DURATION_MS = SESSION_DURATION_SECONDS * 1000
+export const SESSION_INACTIVITY_TIMEOUT_MS = 15 * 60 * 1000 // 15 minutes
+export const SESSION_STORAGE_KEY = '_splunk_rum_sid'
