+++ packages/web/src/session/index.ts
@@ -15,23 +15,5 @@
  * limitations under the License.
  *
  */
-
-import * as assert from 'assert'
-import { generateId } from '../src/utils'
-import { findCookieValue } from '../src/session/cookie-session'
-
-describe('generateId', () => {
-	it('should generate IDs of 64 and 128 bits', () => {
-		const id64 = generateId(64)
-		const id128 = generateId(128)
-		assert.strictEqual(id64.length, 16)
-		assert.strictEqual(id128.length, 32)
-		assert.ok(id64.match('^[0-9a-z]+$'))
-		assert.ok(id128.match('^[0-9a-z]+$'))
-	})
-})
-describe('findCookieValue', () => {
-	it('should not find unset cookie', () => {
-		assert.ok(findCookieValue('nosuchCookie') === undefined)
-	})
-})
+export * from './types'
+export * from './session'
