+++ packages/web/test/SessionBasedSampler.test.ts
@@ -19,10 +19,10 @@
 import * as assert from 'assert'
 import { InternalEventTarget } from '../src/EventTarget'
 import { SessionBasedSampler } from '../src/SessionBasedSampler'
-import { initSessionTracking, updateSessionStatus } from '../src/session'
+import { initSessionTracking, updateSessionStatus } from '../src/session/session'
 import { context, SamplingDecision } from '@opentelemetry/api'
 import { SplunkWebTracerProvider } from '../src'
-import { COOKIE_NAME } from '../src/cookie-session'
+import { COOKIE_NAME } from '../src/session/cookie-session'
 
 describe('Session based sampler', () => {
 	it('decide sampling based on session id and ratio', () => {
