+++ packages/web/test/SplunkOtelWeb.test.ts
@@ -19,7 +19,7 @@
 import { SpanAttributes } from '@opentelemetry/api'
 import { expect } from 'chai'
 import SplunkRum from '../src'
-import { updateSessionStatus } from '../src/session'
+import { updateSessionStatus } from '../src/session/session'
 
 describe('SplunkOtelWeb', () => {
 	afterEach(() => {
